requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.0.3
python-dotenv==1.0.0
